prevent from adding empty for shopping list for phone

+++ phone_app/home/shopping.js
@@ -1,23 +1,24 @@
 customElements.define('add-shopping-item', AddShoppingItemModal)
 
-function addItem(item){
-    var a= item.querySelector("ion-input").value
-    console.log(a)
+function addItem(item) {
+    var a = item.querySelector("ion-input").value
 
-    var toAdd = `<ion-item-sliding>
-                    <ion-item-options side="end">
-                        <ion-item-option color="danger" onclick=deleteItem(this)>
-                            Delete
-                        </ion-item-option>
-                    </ion-item-options>
+    if (a != "") {
+        var toAdd = `<ion-item-sliding>
+        <ion-item-options side="end">
+            <ion-item-option color="danger" onclick=deleteItem(this)>
+                Delete
+            </ion-item-option>
+        </ion-item-options>
 
-                    <ion-item onclick=onShoppingItemClick(this) quantity="0">
-                        <p class="shoppingListName">`+ a + `</p>
-                        <p slot="end" class="shoppingListQuantity"></p>
-                    </ion-item> 
-                    
-                </ion-item-sliding>`
-    item.parentElement.parentElement.querySelector('#shoppingList').innerHTML += toAdd
+        <ion-item onclick=onShoppingItemClick(this) quantity="0">
+            <p class="shoppingListName">` + a + `</p>
+            <p slot="end" class="shoppingListQuantity"></p>
+        </ion-item> 
+        
+    </ion-item-sliding>`
+        item.parentElement.parentElement.querySelector('#shoppingList').innerHTML += toAdd
+    }
 
     item.querySelector("ion-input").value = ""
 }
@@ -33,12 +34,12 @@ function onShoppingItemClick(item) {
 
 async function editShoppingItem(item) {
     const modal = await editShoppingItemModal(item);
-  
-    
+
+
     modal.onDidDismiss().then(data => {
-      if (data["data"]) {
-        console.log("data is: ")
-        console.log(data["data"])
+        if (data["data"]) {
+            console.log("data is: ")
+            console.log(data["data"])
             item.querySelector(".shoppingListName").innerHTML = data["data"]["name"]
             item.attributes.quantity.value = data["data"]["quantity"]
 
@@ -47,13 +48,13 @@ async function editShoppingItem(item) {
             } else {
                 item.querySelector(".shoppingListQuantity").innerHTML = ""
             }
-            
-      }
+
+        }
     })
-  
+
     await modal.present()
 }
-  
+
 async function editShoppingItemModal(item) {
 
     const modalController = document.querySelector('ion-modal-controller');
@@ -71,18 +72,19 @@ async function editShoppingItemModal(item) {
         componentProps: {
             'name': content,
             'quantity': item.attributes.quantity.value
-          }
+        }
     });
 
     return modalElement;
 }
 
-function addClicked(item){
+function addClicked(item) {
     addItem(item)
 }
+
 function enterPressed(e, item) {
     var code = (e.keyCode ? e.keyCode : e.which);
-    if(code == 13) { //Enter keycode
+    if (code == 13) { //Enter keycode
         addItem(item)
     }
 }
